requirements.txt
torch>=2.0.0
torchvision
tqdm
numpy
Pillow
PyYAML
streamlit
